vscode: Get paths from product.json
The VSCode packages contain a product.json which is used to determine
the extension and config dirs. Instead of hardcoding the package name,
this change parses this file (which may appear in one of two places) and
uses that to generate the paths.

This hopefully allows more VSCode derivatives to "just work" without
home-manager updates.